api/wecubek8s/requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
talos-api
dogpile.cache
gunicorn
paramiko
requests
pyjwt<2.0.0
pycrypto
# for platform login encryption, apt install swig
M2Crypto
gevent
pytz
apscheduler
certifi
kubernetes